Added "periodic pair dirichlet" boundary condition option to support pinning rotational degrees of freedom for reflectionally symmetric meshes.
Simulate now prints out the integrated surface tractions for each Dirichlet
region (so you can recover the constraint forces)

msh_processor now supports scalar-scalar binary operations.

+++ ExpressionVector.hh
@@ -41,7 +41,7 @@ struct ExpressionEnvironment {
     }
 
     void setVectorValue(const std::string &name, const std::vector<Real> &v) {
-        for (int i = 0; i < v.size(); ++i)
+        for (size_t i = 0; i < v.size(); ++i)
             setValue(name + std::to_string(i), v[i]);
     }
